Moved everything under the '.chartjs.wicket' super package
