+++ wicket/src/main/java/nl/crashdata/chartjs/wicket/components/panels/SimpleGraphPanel.html
