+++ wicket/src/main/java/nl/crashdata/chartjs/wicket/components/resources/MomentJavaScriptResourceReference.java
@@ -1,4 +1,4 @@
-package nl.crashdata.chartjs.components.resources;
+package nl.crashdata.chartjs.wicket.components.resources;
 
 import org.apache.wicket.request.resource.JavaScriptResourceReference;
