+++ wicket/src/test/java/nl/crashdata/chartjs/wicket/components/pages/SimpleTestPage.html
